Outline of the customer_data.php form, the js for form validation is still incomplete but partially functional.
I have pushed a css file for customer registration as I thought the formats I was applying would be pretty unique to this page, the css can either be added to the main css file when created or kept seperate.

in my comments I made mention of a js and css for a modal for the customer login, I already have this complete from past work and will make some minor changes before pushing to the repository.  If the modal login proves to be too difficult falling back to a dedicated login page is always a possibility.